Create EDA_3_feature_engineering_diversity_regularity_focus_type.ipynb
This script creates multiple new features focused on measuring (1) diversity of types of interaction w/ the VLE, (2) regularity of engagement with the VLE, and (3) the percentage of VLE interactions that were (a) assessment based, (b) content based, or (c) collaboration based. It also creates a column that aggregates all VLE interactions on a single day into one number.

In addition to the main dataframe, the code creates 2 tables. The first table, student_diversity, aggregates the features per student so there is one stat per student for the entire module. (NB: this somewhat misleadingly named--it's not just about diversity of interaction statistics) It also creates a table, week_one, to measure these features *only for the first week*.

Finally, the script uses the week_one table to measure the relationship between these features and final_result.